SBERDOMA-20 Move search.ts to Ticket domain folder
