+++ apps/condo/utils/clientSchema/Ticket/search.ts
@@ -1,9 +1,45 @@
-import {
-    GET_ALL_CLASSIFIERS_QUERY,
-    GET_ALL_ORGANIZATION_EMPLOYEE_QUERY,
-    GET_ALL_PROPERTIES_QUERY,
-    GET_ALL_SOURCES_QUERY,
-} from '../../schema/Ticket.gql'
+const gql = require('graphql-tag')
+
+// TODO(pahaz): add organization filter
+const GET_ALL_SOURCES_QUERY = gql`
+    query selectSource ($value: String) {
+        objs: allTicketSources(where: {name_contains: $value, organization_is_null: true}) {
+            id
+            name
+        }
+    }
+`
+
+// TODO(pahaz): add organization filter
+const GET_ALL_CLASSIFIERS_QUERY = gql`
+    query selectSource ($value: String) {
+        objs: allTicketClassifiers(where: {name_contains: $value, organization_is_null: true, parent_is_null: true}) {
+            id
+            name
+        }
+    }
+`
+
+// TODO(pahaz): add organization filter
+const GET_ALL_PROPERTIES_QUERY = gql`
+    query selectProperty ($value: String) {
+        objs: allProperties(where: {name_contains: $value}) {
+            id
+            name
+        }
+    }
+`
+
+const GET_ALL_ORGANIZATION_EMPLOYEE_QUERY = gql`
+    query selectOrgarnizationEmployee ($value: String, $organization: ID) {
+        objs: allOrganizationEmployees(where: {name_contains: $value, organization: {id: $organization}}) {
+            name
+            user {
+                id
+            }
+        }
+    }
+`
 
 async function _search (client, query, variables) {
     return await client.query({
